requirements.txt
bitsandbytes==0.40.0
peft==0.4.0
accelerate==0.21.0
einops==0.6.1
evaluate==0.4.0
scikit-learn==1.2.2
wandb==0.15.3
pandas
tensorboard
fairscale
sentencepiece
gradio
open_clip_torch
fire
packaging
transformers
pyyaml
pathlib
Ninja
Pillow
opencv-python
tqdm
git+https://github.com/csuhan/timm_0_3_2.git
git+https://github.com/openai/CLIP.git